security test for release-7.1

+++ dbms/src/Flash/Coprocessor/GenSchemaAndColumn.h
@@ -18,7 +18,7 @@
 #include <Core/NamesAndTypes.h>
 #include <Flash/Coprocessor/ChunkCodec.h>
 #include <Flash/Coprocessor/TiDBTableScan.h>
-#include <TiDB/Schema/TiDB.h>
+#include <Storages/Transaction/TiDB.h>
 #include <common/StringRef.h>
 
 namespace DB
@@ -39,6 +39,7 @@ using ColumnDefinesPtr = std::shared_ptr<std::vector<ColumnDefine>>;
 } // namespace DM
 
 // The column defines and `extra table id index`
-std::tuple<DM::ColumnDefinesPtr, int> genColumnDefinesForDisaggregatedRead(const TiDBTableScan & table_scan);
+std::tuple<DM::ColumnDefinesPtr, size_t>
+genColumnDefinesForDisaggregatedRead(const TiDBTableScan & table_scan);
 
 } // namespace DB
